requirements.txt
# AutoCorp Python Agent Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
web3>=6.11.0
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
